+++ internal/service/remote/testhelper_test.go
@@ -1,0 +1,68 @@
+package remote
+
+import (
+	"log"
+	"net"
+	"os"
+	"testing"
+	"time"
+
+	pb "gitlab.com/gitlab-org/gitaly-proto/go"
+	"gitlab.com/gitlab-org/gitaly/internal/rubyserver"
+	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/reflection"
+)
+
+var rubyServer *rubyserver.Server
+
+func TestMain(m *testing.M) {
+	os.Exit(testMain(m))
+}
+
+func testMain(m *testing.M) int {
+	defer testhelper.MustHaveNoChildProcess()
+
+	var err error
+
+	testhelper.ConfigureRuby()
+	rubyServer, err = rubyserver.Start()
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer rubyServer.Stop()
+
+	return m.Run()
+}
+
+func runRemoteServiceServer(t *testing.T) (*grpc.Server, string) {
+	grpcServer := testhelper.NewTestGrpcServer(t, nil, nil)
+	serverSocketPath := testhelper.GetTemporaryGitalySocketFileName()
+
+	listener, err := net.Listen("unix", serverSocketPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	pb.RegisterRemoteServiceServer(grpcServer, &server{rubyServer})
+	reflection.Register(grpcServer)
+
+	go grpcServer.Serve(listener)
+
+	return grpcServer, serverSocketPath
+}
+
+func newRemoteClient(t *testing.T, serverSocketPath string) (pb.RemoteServiceClient, *grpc.ClientConn) {
+	connOpts := []grpc.DialOption{
+		grpc.WithInsecure(),
+		grpc.WithDialer(func(addr string, timeout time.Duration) (net.Conn, error) {
+			return net.DialTimeout("unix", addr, timeout)
+		}),
+	}
+	conn, err := grpc.Dial(serverSocketPath, connOpts...)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	return pb.NewRemoteServiceClient(conn), conn
+}